feat(rabin): try a suboptimal boundary
Instead of to directly break at boundary uplimit, we record a
suboptimal boundary, break, and then rewind.  The boundary may
not exists if the input's entropy is too high, in such a case we
still break on boundary uplimit.

The approach may be costly, but it may offer better dedup results,
and it mirrors the agent's behavior.